+++ packages/headless/src/controllers/commerce/core/facets/regular/headless-commerce-regular-facet.test.ts
@@ -4,12 +4,13 @@ import {
   toggleSelectFacetValue,
 } from '../../../../../features/facets/facet-set/facet-set-actions';
 import {CommerceAppState} from '../../../../../state/commerce-app-state';
-import {MockCommerceEngine, buildMockCommerceEngine} from '../../../../../test';
 import {buildMockCommerceFacetRequest} from '../../../../../test/mock-commerce-facet-request';
 import {buildMockCommerceRegularFacetResponse} from '../../../../../test/mock-commerce-facet-response';
 import {buildMockCommerceFacetSlice} from '../../../../../test/mock-commerce-facet-slice';
 import {buildMockCommerceRegularFacetValue} from '../../../../../test/mock-commerce-facet-value';
 import {buildMockCommerceState} from '../../../../../test/mock-commerce-state';
+import {buildMockCommerceEngine} from '../../../../../test/mock-engine';
+import {MockCommerceEngine} from '../../../../../test/mock-engine';
 import {commonOptions} from '../../../product-listing/facets/headless-product-listing-facet-options';
 import {
   CommerceRegularFacet,