+++ packages/headless/src/controllers/field-suggestions/facet/headless-field-suggestions.test.ts
@@ -1,17 +1,15 @@
-import {SearchAppState} from '../../..';
 import {executeFacetSearch} from '../../../features/facets/facet-search-set/generic/generic-facet-search-actions';
 import {updateFacetSearch} from '../../../features/facets/facet-search-set/specific/specific-facet-search-actions';
 import {registerFacet} from '../../../features/facets/facet-set/facet-set-actions';
 import {defaultFacetOptions} from '../../../features/facets/facet-set/facet-set-slice';
 import {FacetRequest} from '../../../features/facets/facet-set/interfaces/request';
-import {
-  buildMockSearchAppEngine,
-  createMockState,
-  MockSearchEngine,
-} from '../../../test';
+import {SearchAppState} from '../../../state/search-app-state';
+import {MockSearchEngine} from '../../../test/mock-engine';
+import {buildMockSearchAppEngine} from '../../../test/mock-engine';
 import {buildMockFacetRequest} from '../../../test/mock-facet-request';
 import {buildMockFacetSearch} from '../../../test/mock-facet-search';
 import {buildMockFacetSlice} from '../../../test/mock-facet-slice';
+import {createMockState} from '../../../test/mock-state';
 import {
   buildFieldSuggestions,
   FieldSuggestions,