+++ packages/headless/src/features/result-preview/result-preview-actions-loader.ts
@@ -1,6 +1,6 @@
 import {AsyncThunkAction, PayloadAction} from '@reduxjs/toolkit';
-import {CoreEngine} from '../..';
 import {HtmlRequestOptions} from '../../api/search/html/html-request';
+import {CoreEngine} from '../../app/engine';
 import {resultPreviewReducer as resultPreview} from '../../features/result-preview/result-preview-slice';
 import {
   AsyncThunkGlobalOptions,